Update findIndex.md
In this statement, the expression cat === 'Lily' is a single expression, and JavaScript automatically returns the result of this expression. Therefore, there is no need for a return statement here.

If there were more statements in the function or if you were using conditional structures, then you would need to use return. However, in this case, you can achieve the same result without writing return, so it is unnecessary I think